HSEARCH-4104 Add a dedicated Maven profile for Elasticsearch 7.10
