+++ CONTRIBUTING.md
@@ -202,7 +202,8 @@ The following profiles are available:
  * `elasticsearch-7.0` for 7.0 to 7.2
  * `elasticsearch-7.3` for 7.3 to 7.6
  * `elasticsearch-7.7` for 7.7
- * `elasticsearch-7.8` for 7.8+ (the default)
+ * `elasticsearch-7.8` for 7.8 to 7.9
+ * `elasticsearch-7.10` for 7.10+ (the default)
 
 A list of available versions for `test.elasticsearch.connection.version` can be found on
 [Maven Central](https://search.maven.org/search?q=g:org.elasticsearch%20AND%20a:elasticsearch&core=gav).
