+++ parents/integrationtest/pom.xml
@@ -502,9 +502,33 @@
             </build>
         </profile>
 
-        <!-- Elasticsearch 7.8+ test environment (default) -->
+        <!-- Elasticsearch 7.8 to 7.9 test environment -->
         <profile>
             <id>elasticsearch-7.8</id>
+            <properties>
+                <test.elasticsearch.connection.version>7.9.3</test.elasticsearch.connection.version>
+                <test.elasticsearch.testdialect>org.hibernate.search.util.impl.integrationtest.backend.elasticsearch.dialect.Elasticsearch78TestDialect</test.elasticsearch.testdialect>
+            </properties>
+            <build>
+                <pluginManagement>
+                    <plugins>
+                        <plugin>
+                            <groupId>com.github.alexcojocaru</groupId>
+                            <artifactId>elasticsearch-maven-plugin</artifactId>
+                            <version>${version.com.github.alexcojocaru.elasticsearch.plugin}</version>
+                            <configuration>
+                                <pathConf>${project.build.directory}/elasticsearch-maven-plugin/7.8/configuration/</pathConf>
+                                <pathInitScript>${project.build.directory}/elasticsearch-maven-plugin/7.8/init/init.json</pathInitScript>
+                            </configuration>
+                        </plugin>
+                    </plugins>
+                </pluginManagement>
+            </build>
+        </profile>
+
+        <!-- Elasticsearch 7.10+ test environment (default) -->
+        <profile>
+            <id>elasticsearch-7.10</id>
             <activation>
                 <!-- Activate by default, i.e. if test.elasticsearch.connection.version has not been defined explicitly -->
                 <property>
@@ -513,7 +537,7 @@
             </activation>
             <properties>
                 <test.elasticsearch.connection.version>${version.org.elasticsearch.main}</test.elasticsearch.connection.version>
-                <test.elasticsearch.testdialect>org.hibernate.search.util.impl.integrationtest.backend.elasticsearch.dialect.Elasticsearch78TestDialect</test.elasticsearch.testdialect>
+                <test.elasticsearch.testdialect>org.hibernate.search.util.impl.integrationtest.backend.elasticsearch.dialect.Elasticsearch710TestDialect</test.elasticsearch.testdialect>
             </properties>
             <build>
                 <pluginManagement>
